Remove latex_hook.rs — content moved to tex.rs (Perl: TeX.pool.ltxml L33-56)
Autoload triggers (\documentclass, \ExplSyntaxOn, etc.), OmniBus autoloads
(\mathfrak, \theoremstyle, etc.), and file bookkeeping stubs all moved to
tex.rs where they belong per Perl's TeX.pool.ltxml organization.

Three non-Perl engine files now eliminated:
- latex_other_in_appendices.rs → latex_base.rs + latex_constructs.rs
- latex_semi_undocumented.rs → latex_base.rs + latex_constructs.rs
- latex_hook.rs → tex.rs
