+++ latexml_package/src/engine/tex.rs
@@ -1,4 +1,29 @@
 use crate::prelude::*;
+
+/// Perl: DefAutoload — define a macro that auto-loads a package on first use.
+/// When the command is first invoked, it loads the specified package (via RequirePackage),
+/// then re-emits the original CS so it gets re-executed with the proper definition.
+fn def_autoload(cs_name: &str, package: &str) -> Result<()> {
+  use latexml_core::definition::ExpansionBody;
+  let cs_tok = T_CS!(cs_name);
+  // Don't overwrite if already defined
+  if IsDefined!(&cs_tok) {
+    return Ok(());
+  }
+  let pkg_name = package.to_string();
+  let cs_for_closure = cs_tok;
+  def_macro(
+    cs_tok,
+    None,
+    ExpansionBody::Closure(Rc::new(move |_args| {
+      require_package(&pkg_name, RequireOptions::default())?;
+      Ok(Tokens::new(vec![cs_for_closure]))
+    })),
+    None,
+  )?;
+  Ok(())
+}
+
 LoadDefinitions!({
   // port of TeX.pool.ltxml
   // commit 4cd73e7584c5f0422293ba38f9b757332584afec
@@ -45,7 +70,93 @@ LoadDefinitions!({
   DefRegister!("\\everyhelp", Tokens!());
   DefMacro!("\\hiderel{}", "#1"); // Just ignore, for now...
 
-  InnerPool!(latex_hook);
+  // Perl TeX.pool.ltxml L33-56: autoload triggers for LaTeX, expl3, AmSTeX
+  // (moved from latex_hook.rs which has no Perl equivalent)
+  for ltxtrigger in [
+    "\\documentclass",
+    "\\newcommand",
+    "\\renewcommand",
+    "\\newenvironment",
+    "\\renewenvironment",
+    "\\NeedsTeXFormat",
+    "\\ProvidesPackage",
+    "\\RequirePackage",
+    "\\ProvidesFile",
+    "\\makeatletter",
+    "\\makeatother",
+    "\\begin",
+    "\\listfiles",
+    "\\nofiles",
+    "\\typeout",
+    "\\PassOptionsToPackage",
+  ]
+  .iter()
+  {
+    DefMacro!(T_CS!(ltxtrigger), None, {
+      Tokens!(T_CS!("\\@load@latex@pool"), T_CS!(ltxtrigger))
+    });
+  }
+  DefPrimitive!("\\@load@latex@pool", {
+    input_definitions("LaTeX", InputDefinitionOptions {
+      extension: Some(Cow::Borrowed("pool")),
+      ..InputDefinitionOptions::default()
+    })?;
+  });
+
+  // Perl TeX.pool.ltxml L42-48: expl3 triggers
+  for ltx3trigger in [
+    "\\ExplSyntaxOn",
+    "\\ProvidesExplClass",
+    "\\ProvidesExplPackage",
+  ] {
+    def_autoload(ltx3trigger, "expl3")?;
+  }
+
+  // OmniBus autoloads (Perl: OmniBus.cls.ltxml DefAutoload entries)
+  def_autoload("\\mathfrak", "amsfonts")?;
+  def_autoload("\\mathbb", "amsfonts")?;
+  def_autoload("\\Bbb", "amsfonts")?;
+  def_autoload("\\theoremstyle", "amsthm")?;
+  def_autoload("\\numberwithin", "amsmath")?;
+  def_autoload("\\align", "amsmath")?;
+  def_autoload("\\subequations", "amsmath")?;
+  def_autoload("\\multline", "amsmath")?;
+  def_autoload("\\curraddr", "ams_support")?;
+  def_autoload("\\subjclass", "ams_support")?;
+
+  // File bookkeeping (Perl TeX.pool.ltxml, needed before LaTeX.pool loads)
+  DefMacro!(
+    "\\@pushfilename",
+    r"\xdef\@currnamestack{{\@currname}{\@currext}{\the\catcode`\@}\@currnamestack}"
+  );
+  DefMacro!(
+    "\\@popfilename",
+    r"\expandafter\@p@pfilename\@currnamestack\@nil"
+  );
+  DefMacro!(
+    "\\@p@pfilename {}{}{} Until:\\@nil",
+    r"\gdef\@currname{#1}%
+      \gdef\@currext{#2}%
+      \catcode`\@#3\relax
+      \gdef\@currnamestack{#4}"
+  );
+  DefMacro!(T_CS!("\\@currnamestack"), None, Tokens!());
+  Let!("\\@currname", "\\lx@empty");
+  Let!("\\@currext", "\\lx@empty");
+
+  // Early stubs needed by ProcessOptions/DeclareOption before LaTeX.pool loads
+  if !IsDefined!(&T_CS!("\\@unknownoptionerror")) {
+    DefPrimitive!("\\@unknownoptionerror", {});
+  }
+  if !IsDefined!(&T_CS!("\\OptionNotUsed")) {
+    DefPrimitive!("\\OptionNotUsed", {});
+  }
+  if !IsDefined!(&T_CS!("\\AtBeginDocument")) {
+    DefMacro!("\\AtBeginDocument{}", "");
+  }
+  if !IsDefined!(&T_CS!("\\@addtofilelist")) {
+    DefMacro!("\\@addtofilelist{}", "");
+  }
 
   // Perl: LoadFormat('plain') — precompiled plain.tex state.
   // TODO: Enable once dump has full parity (Let, CharDef, Register entries).
